Ensuring indexes are int not float
Using casting to turn calculations for index of list from float to integer

+++ median.py
@@ -6,11 +6,11 @@ while True:
         print("Enter a list of numbers separated by commas: ")
         numbers = [float(value) for value in input().split(",")]
         if len(numbers)%2==0:
-            middle1=numbers[len(numbers)/2]
-            middle2=numbers[(len(numbers)/2)-1]
+            middle1=numbers[int(len(numbers)/2)]
+            middle2=numbers[int((len(numbers)/2)-1)]
             median = (middle1+middle2)/2
         else:
-            median = numbers[(len(numbers)-1)/2]
+            median = numbers[int((len(numbers)-1)/2)]
     except ValueError:
         print("Some input could not be converted to a number!")
     else:
